+++ packages/dashpub_cli/lib/src/commands/token_command.dart
@@ -28,7 +28,10 @@ class TokenCommand extends Command {
 
   @override
   Future<void> run() async {
-    final url = argResults!['url'] as String;
+    var url = argResults!['url'] as String;
+    if (url.endsWith('/')) {
+      url = url.substring(0, url.length - 1);
+    }
     final generate = argResults!['generate'] as bool;
     final config = DashpubConfig();
     final currentToken = await config.getToken();